003 One to One querying relationship

+++ database/migrations/2023_12_24_065203_create_profiles_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('profiles', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->unsignedInteger('author_id')->unique();
+            $table->unsignedBigInteger('author_id')->unique();
             $table->foreign('author_id')->references('id')->on('authors');
         });
     }
